Discover and verify samples embedded in asciidoctor sources
This change allows users to test samples embedded in Asciidoctor.

SampleDiscovery.embeddedSamples() looks within all .adoc and
.asciidoc files for blocks marked with the .testable-sample role
and extracts source blocks with a file name to a temp directory,
then extracts a terminal command beginning with "$ " and its
output and verifies that the given sources + command produce the
given output, with normalization of course.

This is one solution for issue #2.