Simplify usage of Sample rule
Since JUnit does not guarantee the order in which rules are evaluated,
using `into(TemporaryFolder)` was error-prone. The only way to make it
safe is to use `RuleChain.outerRule(temporaryFolder).around(sample)`
which is quite involved.

Thus, this commit makes it simpler to use the `Sample` rule when the
used folder is irrelevant. By default, the rule will not create its own
temporary folder (using `TemporaryFolder` internally). To make it more
explicit or to specify a custom parent directory, users may call
`intoTemporaryFolder()` or `intoTemporaryFolder(parentDir)`,
respectively.